+++ scripts/new_j_phys_22_063041 - 2a.py
@@ -35,7 +35,6 @@ params = {
     'plotter': {
         'type': 'pcolormesh',
         'palette': 'Greens',
-        'bins': 11,
         'title': '$G = g = 3 \\kappa$',
         'x_label': '$\\Delta / \\omega_{m}$',
         'x_ticks': [0.4, 1.0, 1.6],