+++ scripts/opt_lett_41_2676 - 2.py
@@ -27,13 +27,10 @@ params = {
     'plotter': {
         'type': 'surface_cz',
         'x_label': '$x$',
-        'x_bound': 'both',
         'x_ticks': [0, 75, 150],
         'y_label': '$\\tau$',
-        'y_bound': 'both',
         'y_ticks': [0, 1, 2],
         'v_label': '$|\\alpha| (10^{3})$',
-        'v_bound': 'both',
         'v_ticks': [0, 1000, 2000],
         'v_tick_labels': [0, 1, 2],
         'show_cbar': False,
@@ -45,7 +42,7 @@ params = {
 init_log()
 
 # initialize system
-system = OptLett41_2676(params['system'])
+system = OptLett41_2676(params=params['system'])
 
 # get mode amplitude dynamics
 amps, T, X = system.get_mode_amplitude_dynamics(solver_params=params['solver'], plot=True, plotter_params=params['plotter'])